test(conftest): fix teardown error

+++ tests/compare/conftest.py
@@ -86,5 +86,6 @@ def compare_ndarrays(ndarrays_regression: NDArraysRegressionFixture):
 
     yield kernel
 
-    if not any(source_data_dir.iterdir()): # type: ignore
-        source_data_dir.rmdir() # type: ignore
+    if source_data_dir.exists(): # type: ignore
+        if not any(source_data_dir.iterdir()): # type: ignore
+            source_data_dir.rmdir() # type: ignore
